split pruner provider workspace into focused components
